### 2020-03-10 #2159 @whedon
[PRE REVIEW]: MPA: a novel cross-language API for time series analysis
**Submitting author:** @vanbenschoten (<a href="http://orcid.org/0000-0002-7944-6237">Andrew Van Benschoten</a>)
**Repository:** <a href="https://github.com/matrix-profile-foundation/joss_paper" target ="_blank">https://github.com/matrix-profile-foundation/joss_paper</a>
**Version:** Misc (multiple repos)
