### 2020-03-10 #2159 @vanbenschoten
@danielskatz there are three individual repos that comprise the software described in our paper. The manuscript is in matrix-profile-foundation/joss_paper, and the actual codebases are:

1. matrix-profile-foundation/tsmp
2. matrix-profile-foundation/go-matrixprofile
3. matrix-profile-foundation/matrixprofile

I connected with @arfon last week about how to handle this situation, and he advised me that submitting the manuscript in a separate repo was the recommended approach.

Let me know if there are any further questions!
