### 2020-03-18 #2159 @arfon
@suyashkumar, @robohyun66, @galenstocking, @ionlights - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
