### 2020-04-05 #2159 @arfon
@bliiir @yzhao062 @Sirjla @krystophny @coljac - - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

This review would need you to inspect code in a [small number of repositories](https://github.com/openjournals/joss-reviews/issues/2159#issuecomment-597276663):

- https://github.com/matrix-profile-foundation/tsmp
- https://github.com/matrix-profile-foundation/go-matrixprofile
- https://github.com/matrix-profile-foundation/matrixprofile

...and the paper explaining how they relate is available [here](https://github.com/openjournals/joss-papers/blob/joss.02159/joss.02159/10.21105.joss.02159.pdf).
