### 2020-04-22 #2159 @vanbenschoten
Hi Arfon,
   Any update on the publication front? We're getting to the point where we
need to DOI to continue other lines of work. Thanks!

-Andrew
