### 2020-04-22 #2159 @arfon
> Any update on the publication front?

I'm afraid not sorry - it looks like we're really struggling to get any traction here with potential reviewers. This could be due to people struggling to keep up with life/their day jobs during the pandemic, or lack of interest/availability to review. Let me try another couple of folks here...
