### 2020-04-22 #2159 @coljac
I'll be take this one on if the editors would like. I didn't pipe up before as I'm not an R user, but I'll figure that part of it out.
