### 2020-04-22 #2159 @jkahn
Not a good one for me; R is far enough out of my wheelhouse to make it
difficult for me to compare to existing work.

On Wed, Apr 22, 2020, 4:25 PM Andrew Van Benschoten <
notifications@github.com> wrote:

> @coljac thanks so much! We really appreciate it!
>
> On Wed, Apr 22, 2020, 5:47 PM Colin Jacobs <notifications@github.com>
> wrote:
>
> > I'll be take this one on if the editors would like. I didn't pipe up
> > before as I'm not an R user, but I'll figure that part of it out.
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <
> https://github.com/openjournals/joss-reviews/issues/2159#issuecomment-618078877
> >,
> > or unsubscribe
> > <
> https://github.com/notifications/unsubscribe-auth/AB53ISHDQP3Z4SGM2E47DF3RN5XXBANCNFSM4LFGIYNQ
> >
> > .
> >
>
