### 2020-04-22 #2159 @arfon
> Not a good one for me; R is far enough out of my wheelhouse to make it
> difficult for me to compare to existing work.

Got it, thanks for responding @jkahn

> I'll be willing take this one on if the editors would like. I didn't pipe up before as I'm not an R user, but I'll figure that part of it out.

@coljac - many thanks! I'll add you as a reviewer now but won't start the actual review process until we have identified a second reviewer.
