### 2020-04-23 #2179 @whedon
[REVIEW]: MPA: a novel cross-language API for time series analysis
**Submitting author:** @vanbenschoten (<a href="http://orcid.org/0000-0002-7944-6237">Andrew Van Benschoten</a>)
**Repository:** <a href="https://github.com/matrix-profile-foundation/joss_paper" target ="_blank">https://github.com/matrix-profile-foundation/joss_paper</a>
**Version:** Misc (multiple repos)
