### 2020-04-23 #2179 @arfon
@coljac, @krystophny - thanks again for agreeing to review this submission! Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Unlike some JOSS submissions, the code for this review is in a [small number of repositories](https://github.com/openjournals/joss-reviews/issues/2159#issuecomment-597276663):

- https://github.com/matrix-profile-foundation/tsmp
- https://github.com/matrix-profile-foundation/go-matrixprofile
- https://github.com/matrix-profile-foundation/matrixprofile
