### 2020-04-23 matrix-profile-foundation/joss_paper#1 @krystophny
Clarify location of source code
@vanbenschoten I have the point "Repository: Is the source code for this software available at the repository url?" on my reviewer checklist. It is not, since you created a separate repository for the paper only. Could you clarify how this should be handled for JOSS? It's my first review so I have no idea.
