### 2020-04-23 matrix-profile-foundation/joss_paper#1 @tylerwmarrs
@krystophny please see the message that arfon added. This paper is associated with three other libraries. I hope this helps. https://github.com/openjournals/joss-reviews/issues/2179#issuecomment-618374612
