### 2020-04-23 matrix-profile-foundation/joss_paper#1 @krystophny
@tylerwmarrs thanks, everything clear.
