### 2020-04-23 matrix-profile-foundation/joss_paper#2 @krystophny
Contribution and authorship
@vanbenschoten I see only a single commit from you in the three repositories for this paper:

* https://github.com/matrix-profile-foundation/tsmp (mainly @franzbischoff )
* https://github.com/matrix-profile-foundation/go-matrixprofile (mainly @aouyang1 )
* https://github.com/matrix-profile-foundation/matrixprofile (mainly @tylerwmarrs , some @nikita-smyrnov )

Please clarify your contribution as the main author.
