### 2020-04-23 matrix-profile-foundation/joss_paper#2 @vanbenschoten
@krystophny Definitely! My role has primarily been that of the Primary Investigator/Corresponding Author. We decided to list names alphabetically given that we see all of our contributions as equal, but I'm also happy to list my name in the more traditional corresponding author spot at the end.

1) I'm the head of the Matrix Profile Foundation and brought this great group of folks together to form the MPF body & vision.

2) I'm the original author of the first Python implementation of Matrix Profile (https://github.com/target/matrixprofile-ts), which helped shape our viewpoint into how the libraries should be designed and evolve.

2) I was heavily involved in architecting the Matrix Profile API itself.

3) I'm the author of the JOSS paper
