### 2020-04-23 matrix-profile-foundation/joss_paper#2 @krystophny
@vanbenschoten thanks for the info. Just one more question: why not include https://github.com/target/matrixprofile-ts in the submission? Has this already been published elsewhere, or what's the reason not to include it?
