### 2020-04-23 matrix-profile-foundation/joss_paper#2 @vanbenschoten
@krystophny We didn't include matrixprofile-ts because Target (who still owns the GitHub repo) decided they didn't want the codebase to be part of the broader body of MPF work.
