### 2020-04-23 matrix-profile-foundation/joss_paper#2 @krystophny
Alright. Based on this information I can conclude that besides formulating the paper you did major conceptual developments and wrote the paper, while your original Python code is not part of the contribution for organizational reasons. It has however been used as a template for the other codes. This should justify your authorship.
