### 2020-04-23 #2179 @krystophny
### General checks

1) The source code of all three mentioned repositories is available. Please check that this information (links to the URLs) appears also in the final publication https://github.com/matrix-profile-foundation/joss_paper/issues/1
2) The codes are under the Apache License, which is OSI compliant and LICENSE files exist
3) All authors appear to have contributed significantly, the list is alphabetical https://github.com/matrix-profile-foundation/joss_paper/issues/2
