+++ paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 # Summary
 
-Two fundamental tasks in time series analysis are identifying anomalous events (“discords”) and repeated patterns (“motifs”). Successfully accomplishing these tasks is of the utmost importance across many disciplines, and can lead to powerful technological advancements, prevention of catastrophic failures and the generation of significant economic gain. While there are dozens of existing approaches for finding motifs and discords, they are hampered by a combination of steep methodological learning curves, numerous parameters that require tuning and the inability to scale across large datasets [@MP1]. The explosive growth of the data science community provides an additional hurdle for traditional time series analysis methods, as many practitioners lack experience in advanced mathematical and statistical principles. In this paper we present MPA (the _Matrix Profile API_) as a solution to all of these challenges. MPA is a cross-language platform in Python (_matrixprofile_), R (_tsmp_) and Golang (_go-matrixprofile_) that leverages a novel data transformation known as the Matrix Profile [@MP1] to rapidly identify motifs and discords. Perhaps most importantly, MPA is an easy-to-use API that’s relevant for time series novices and experts alike.
+Two fundamental tasks in time series analysis are identifying anomalous events (“discords”) and repeated patterns (“motifs”). Successfully accomplishing these tasks is of the utmost importance across many disciplines, and can lead to powerful technological advancements, prevention of catastrophic failures and the generation of significant economic gain. While there are dozens of existing approaches for finding motifs and discords, they are hampered by a combination of steep methodological learning curves, numerous parameters that require tuning and the inability to scale across large datasets [@MP1]. The explosive growth of the data science community provides an additional hurdle for traditional time series analysis methods, as many practitioners lack experience in advanced mathematical and statistical principles. In this paper we present MPA (the _Matrix Profile API_) as a solution to all of these challenges. MPA is a cross-language platform in Python ([_matrixprofile_](https://github.com/matrix-profile-foundation/matrixprofile)), R ([_tsmp_](https://github.com/matrix-profile-foundation/tsmp)) and Golang ([_go-matrixprofile_](https://github.com/matrix-profile-foundation/go-matrixprofile])) that leverages a novel data transformation known as the Matrix Profile [@MP1] to rapidly identify motifs and discords. Perhaps most importantly, MPA is an easy-to-use API that’s relevant for time series novices and experts alike.
 
 The intuition behind Matrix Profile is straightforward. It begins with a snippet of data and then slides it along the rest of the time series, calculating the overlap at each new position. More specifically, it evaluates the Euclidean distance between a subsequence and every possible time series segment of the same length, building up the snippet’s “Distance Profile.” If the subsequence repeats itself in the data, there will be at least one perfect match and the minimum Euclidean distance will be zero, or close to zero in the presence of noise. In contrast, if the subsequence is highly unique due to the presence of outliers, matches will be poor and all overlap scores will be high. Every possible snippet is slid across the time series, building up a collection of Distance Profiles. The minimum value for each time step across all distance profiles is collected, creating the time series' final Matrix Profile. Both ends of the Matrix Profile value spectrum are useful. High values indicate uncommon patterns or anomalous events; in contrast, low values highlight repeatable motifs.
 
