### 2020-04-23 matrix-profile-foundation/joss_paper#1 @tylerwmarrs
@krystophny - For clarity, I linked the libraries mentioned in the paper directly to their respective GitHub repositories.
