### 2020-04-29 #2179 @coljac
Hi @vanbenschoten:

Congratulations of developing Matrix profile and the associated software packages. A lot of work has gone into making the algorithms as accessible and usable as possible. This package certainly clears the bar for significance and quality.

I have reviewed the python and golang implementations of matrix profile. The python package is fully-fledged and well documented, with many accessible examples and quality API documentation. The performance was good even with large data sets that I tested. I was able to successfully run all tests.

The golang implementation is much less fleshed out. While a few examples were provided, it lacks proper API documentation. On its own, I'm not sure that go-matrixprofile would clear the bar for this reason. However, the golang implementation is provided alongside two other repositories where the capabilities of the algorithm are more clearly explained. I will discuss this with my co-reviewer, @krystophny; if the R version is similarly well fleshed out to the python version, this may not be a barrier to acceptance.

Regarding the paper, there isn't much discussion of other approaches to the analysis of time series data. The paper would be improved if you named a few other approaches, with references, in the first paragraph (where you have cited Yeh+ 2016).

*Minor issues:*
- The link to example code in the go-matrixprofile README.md is broken
- Python matrix profile: the Linux/Mac instructions to install from source/run tests:
    - the `git clone git@github.com:...` gives a permission error; maybe use https:// URL?
    - The `pip install -e .` fails unless it's run after the `pip install -r requirements.txt` for me.
