+++ paper.bib
@@ -6,6 +6,32 @@
 	note={\url{http://www.cs.unm.edu/~mueen/FastestSimilaritySearch.html}}
 }
 
+@inproceedings{REF1, 
+	place={New York, USA},
+	title={{Anomaly Detection in Univariate Time Series: An Empirical Comparison of Machine Learning Algorithms}},
+	booktitle={{Industrial Conference on Data Mining (ICDM)}},
+	publisher={IEEE},
+	author={Däubener, Sina and Schmitt, Sebastian and Wang, Hao and Bäck, Thomas and Krause, Peter},
+	year={2019},
+	month={Dec}
+}
+
+@misc{REF2,
+    title={Real-Time Anomaly Detection for Streaming Analytics},
+    author={Subutai Ahmad and Scott Purdy},
+    year={2016},
+    eprint={1607.02480},
+    archivePrefix={arXiv},
+    primaryClass={cs.AI}
+}
+
+@Book{REF3,
+    title={Real-Time Anomaly Detection for Streaming Analytics},
+    author={C.M. Bishop},
+    year={2006},
+    publisher={Information Science and Statistics, Springer, New York}
+}
+
 @inproceedings{MP1, 
 	place={Barcelona, Spain},
 	title={{Matrix Profile I: All Pairs Similarity Joins for Time Series: A Unifying View That Includes Motifs, Discords and Shapelets}},
