### 2020-05-02 #2179 @vanbenschoten
@coljac @krystophny

We've updated the paper intro per @coljac 's comments, as well as resolved the matrixprofile install and Golang README issues.
