### 2020-05-02 #2179 @coljac
@vanbenschoten This resolves my issues with the paper.

Once @krystophny has had a chance to look at the R module, we'll get back to you again.
