### 2020-05-04 #2179 @krystophny
@vanbenschoten I could now test installation and basic functionality of your packages in Python, R and Go. There was one issue with the Python version that downgraded my SciPy from 1.4.1 to some 1.3 version. It would be great if you could fix that before publication, or at least prioritize this issue, as it will destroy other dependencies (e.g. tensorflow 2.2).

Otherwise I'm happy with usability and documentation and think the software is very useful. The reference documentation for the R version API is also very detailed and available on CRAN too, so this should be fine @coljac . The paper is well written and explains the basics of what's going on in a concise way.

The last remaining issue are references. JOSS explicitly asks for referencing related codes in the paper. I saw this on the github READMEs only though, and even that one was missing e.g. this code: https://github.com/TDAmeritrade/stumpy that I found when quickly googling for matrix profile in Python. Please give some information where your code stands, what distinguishes it from the others, and also reference the most important alternatives.
