### 2020-05-05 #2179 @vanbenschoten
@krystophny sounds good! I'll add a paragraph or two detailing the history and current landscape of matrix profile implementations.
