### 2020-05-05 #2179 @coljac
@vanbenschoten @krystophny Unless I missed something, I'm not sure about the scipy versioning issue being a problem here. It's standard these days to run any software like this in its own environment to manage these sorts of versioning clashes. Perhaps the docs should reference this, but from where I sit this is understood implicitly.
