### 2020-05-05 #2179 @vanbenschoten
We've gone ahead and resolved the scipy version issue in a minor code
release we're pushing tomorrow, so I think we'll be good on this front.
