### 2020-05-05 #2179 @vanbenschoten
@krystophny @coljac @arfon I've updated the paper to per @krystophny 's feedback.
