### 2020-05-05 #2179 @krystophny
@vanbenschoten very nice work indeed! With the changes made I can propose to accept the paper.

PS: In the proofs I saw a space is missing after "UCR". Is "pan-Matrix Profile" really written in that way?
