### 2020-05-05 #2179 @arfon
Thanks @krystophny and @coljac!

@vanbenschoten - I think this is ready to move forward to the accept/publish stage.

At this point could you make new releases of these software packages review here that includes the changes that have resulted from this review. Then, please make an archive of the software packages together in a deposit with Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
